Fixes concurrent access issues using a concurrent dictionary
