+++ LiteDB/Client/Mapper/BsonMapper.Deserialize.cs
@@ -213,6 +213,7 @@ namespace LiteDB
                 }
 
                 var entity = this.GetEntityMapper(type);
+                entity.WaitForInitialization();
 
                 // initialize CreateInstance
                 if (entity.CreateInstance == null)
