+++ LiteDB/Client/Mapper/EntityBuilder.cs
@@ -30,6 +30,7 @@ namespace LiteDB
         {
             return this.GetMember(member, (p) =>
             {
+                _entity.WaitForInitialization();
                 _entity.Members.Remove(p);
             });
         }
@@ -54,9 +55,11 @@ namespace LiteDB
         {
             return this.GetMember(member, (p) =>
             {
+                _entity.WaitForInitialization();
+                
                 // if contains another _id, remove-it
                 var oldId = _entity.Members.FirstOrDefault(x => x.FieldName == "_id");
-
+        
                 if (oldId != null)
                 {
                     oldId.FieldName = _mapper.ResolveFieldName(oldId.MemberName);
@@ -73,6 +76,7 @@ namespace LiteDB
         /// </summary>
         public EntityBuilder<T> Ctor(Func<BsonDocument, T> createInstance)
         {
+            _entity.WaitForInitialization();
             _entity.CreateInstance = v => createInstance(v);
 
             return this;
@@ -95,7 +99,8 @@ namespace LiteDB
         private EntityBuilder<T> GetMember<TK, K>(Expression<Func<TK, K>> member, Action<MemberMapper> action)
         {
             if (member == null) throw new ArgumentNullException(nameof(member));
-
+            _entity.WaitForInitialization();
+            
             var memb = _entity.GetMember(member);
 
             if (memb == null)
