+++ LiteDB/Client/Mapper/Linq/LinqExpressionVisitor.cs
@@ -619,6 +619,7 @@ namespace LiteDB
 
             // get class entity from mapper
             var entity = _mapper.GetEntityMapper(member.DeclaringType);
+            entity.WaitForInitialization();
 
             // get mapped field from entity
             var field = entity.Members.FirstOrDefault(x => x.MemberName == name);
