+++ LiteDB/Utils/LiteException.cs
@@ -57,6 +57,7 @@ namespace LiteDB
         public const int NOT_ENCRYPTED = 216;
         public const int INVALID_PASSWORD = 217;
         public const int ILLEGAL_DESERIALIZATION_TYPE = 218;
+        public const int ENTITY_INITIALIZATION_FAILED = 219;
 
         public const int INVALID_DATAFILE_STATE = 999;
